Do not send gcc commands involving -Wa, followed by a file name
The Linux kernel since 3.17 does this :

  CODE16GCC_CFLAGS := -m32 -Wa,$(srctree)/arch/x86/boot/code16gcc.h

This makes the asm load a local file. We don't want to send this
to a remote as it fails. In fact whatever follows "-Wa," is
directly passed to the assembler, so either it starts with a "-"
and it's an option, or it does *not* start with a "-" and it's a
file. This patch ensures that anything not starting with a "-" is
not sent remotely.